api/requirements.txt
azure-functions
azure-functions-durable
openai
praat-parselmouth
numpy
azure-ai-transcription==1.0.0b3
azure-cosmos
azure-storage-blob
